Create MooseBot 3.1 - 2 Drivers, 1 Robot
This transfers the flywheel speed adjustments (adjust settings, not switch settings) to a second joystick (channel 5). This also means that if we ever add light sensors to the robot to detect bonus & regular balls, we can use the 4-button channels to tell the robot whether a ball made it into hte high, low, or no goal fr self-scoring purposes.